start love and hate state + reducer
